refactor(cli): delete legacy useReactToolScheduler and rename useToolExecutionScheduler to useToolScheduler
